Move events to src root
