+++ lib/src/main/java/com/zetcode/events/AnonymousInnerClassEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.events;
 
 import javax.swing.GroupLayout;
 import javax.swing.JButton;
@@ -8,20 +8,28 @@ import java.awt.EventQueue;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class DerivedClassEx extends JFrame {
+public class AnonymousInnerClassEx extends JFrame {
 
-    public DerivedClassEx() {
+    public AnonymousInnerClassEx() {
 
         initUI();
     }
 
     private void initUI() {
 
-        var closeBtn = new MyButton("Close");
+        var closeBtn = new JButton("Close");
+
+        closeBtn.addActionListener(new ActionListener() {
+            
+            @Override
+            public void actionPerformed(ActionEvent event) {
+                System.exit(0);
+            }
+        });
 
         createLayout(closeBtn);
 
-        setTitle("Derived class");
+        setTitle("Anonymous inner class");
         setLocationRelativeTo(null);
         setDefaultCloseOperation(EXIT_ON_CLOSE);
     }
@@ -33,6 +41,7 @@ public class DerivedClassEx extends JFrame {
         pane.setLayout(gl);
 
         gl.setAutoCreateContainerGaps(true);
+        gl.setAutoCreateGaps(true);
 
         gl.setHorizontalGroup(gl.createSequentialGroup()
                 .addComponent(arg[0])
@@ -47,25 +56,11 @@ public class DerivedClassEx extends JFrame {
         pack();
     }
 
-    private class MyButton extends JButton implements ActionListener {
-
-        public MyButton(String text) {
-
-            super.setText(text);
-            addActionListener(this);
-        }
-
-        @Override
-        public void actionPerformed(ActionEvent e) {
-            System.exit(0);
-        }
-    }
-
     public static void main(String[] args) {
 
         EventQueue.invokeLater(() -> {
             
-            var ex = new DerivedClassEx();
+            var ex = new AnonymousInnerClassEx();
             ex.setVisible(true);
         });
     }
