+++ lib/src/main/java/com/zetcode/events/MultipleListenersEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.events;
 
 import javax.swing.BorderFactory;
 import javax.swing.GroupLayout;
